Change flake8 options, and move to config file
flake8 is configured to always exit with an exit code of 0. This seems
pointless: developers are likely to check CI output only when an error
occurs, so why bother running flake8 and then making it report success
regardless of what it finds?

Drop the McCabe cyclomatic complexity check. It's a useful metric of
complexity, but affinity is currently very fresh and rough, and I'd like
us to concentrate more on high-level design issues like "how will
affinity be used by tests" and "might it be useful for a Mesh to be
usable as a set or a dict?"

Move the flake8 configuration options to a config file. This makes it
more convenient for developers to run flake8 with the same options that
are used by CI.