Transition from setuptools to poetry (#15)
The QE team is considering switching from setuptools to poetry for a
collection of projects.  It's a good idea to test out this tooling
change with a small, reasonably simple project like this, as it'll give
us a chance to find and work out kinks, or discover the infeasibility of
using poetry before undertaking a much larger effort.