Depend upon receptor from source, not pypi
The version of receptor that's currently uploaded to pypi is a bit
outdated. Notably, it lacks the `-d` option, as in the following:

    receptor -d /my/data/dir …

Receptor is currently being developed at an extremely rapid page, with
little-to-no history of stable versions available, so it makes sense to
depend upon the bleeding-edge version. I expect this to change as the
project becomes more established.