Relax setuptools version constraint
When receptor's tests are executed on my host, the following error
appears:

    pkg_resources.DistributionNotFound: The 'setuptools<46.0.0,>=45.1.0' distribution was not found and is required by receptor-affinity

This might be related to the fact that my host has setuptools 44.0.0
installed system-wide. I suspect that poetry fails to recognize that it
should install an older version of setuptools into a virtualenv so that
receptor's requirements may be satisfied, and instead falls back to
using the system-wide package. I'm not sure.

Ultimately, more research needs to be done so that the source of  this
error can be tracked down, and a proper resolution found.

In the meantime, this commit simply drops version constraints for
setuptools. This should work because we use very little from the
library:

```console
$ git grep pkg_resources receptor_affinity
receptor_affinity/__init__.py:from pkg_resources import get_distribution, DistributionNotFound
```