Change how validation is performed (#13)
* Depend upon receptor from source, not pypi

The version of receptor that's currently uploaded to pypi is a bit
outdated. Notably, it lacks the `-d` option, as in the following:

    receptor -d /my/data/dir …

Receptor is currently being developed at an extremely rapid page, with
little-to-no history of stable versions available, so it makes sense to
depend upon the bleeding-edge version. I expect this to change as the
project becomes more established.

* Relax setuptools version constraint

When receptor's tests are executed on my host, the following error
appears:

    pkg_resources.DistributionNotFound: The 'setuptools<46.0.0,>=45.1.0' distribution was not found and is required by receptor-affinity

This might be related to the fact that my host has setuptools 44.0.0
installed system-wide. I suspect that poetry fails to recognize that it
should install an older version of setuptools into a virtualenv so that
receptor's requirements may be satisfied, and instead falls back to
using the system-wide package. I'm not sure.

Ultimately, more research needs to be done so that the source of  this
error can be tracked down, and a proper resolution found.

In the meantime, this commit simply drops version constraints for
setuptools. This should work because we use very little from the
library:

```console
$ git grep pkg_resources receptor_affinity
receptor_affinity/__init__.py:from pkg_resources import get_distribution, DistributionNotFound
```

* Change how validation is performed

In my opinion, the affinity framework is overly verbose. It prints
messages to stdout regardless of whether tests are going well or badly.
This drastically lowers the signal to noise ratio.

Change this situation by rewriting the `validate_*` methods on the node
and mesh classes.

Currently, these methods print a message about the inspection process
every time they execute. They're also quite conservative about making
callers aware of errors, merely returning a boolean instead of something
more attention-grabbing when something goes wrong. This is reminiscent
of how C programs return codes and hope that the caller is
on-point-enough to remember to check that code.

Make the following changes:

*   Silence the `validate_*` methods, i.e. do not print to stdout,
    stderr, or really anything, ever. It's the job of the test suite to
    generate output as necessary.
*   Throw exceptions instead of returning booleans.

For an example of how this plays out, consider the following test
method:

```python
def test_default_routes_validate(random_mesh):
    assert random_mesh.validate_all_node_routes()
```

This method could be rewritten like so:

```python
def test_default_routes_validate(random_mesh):
    random_mesh.validate_routes()
```

If this test fails, then the test framework will be able to present
more-useful information to the human reviewing the test failure, as the
test won't have a failure message like "False != True," but will instead
have whatever rich error message the developers put into the exception.

Also note that there is no longer a need to use "assert." This is a
minor but rather useful change. Consider the following test method:

```python
def test_pings_perf(mesh):
    results = mesh.ping()
    mesh.validate_ping_results(results)
```

As far as I can tell, the author of this test forgot to add an "assert"
statement into the test, meaning that it will always pass. If all
`validate_*` methods are rewritten in this style, such omissions will no
longer be problematic.

Finally, note that test code like the following also changes:

```python
wait_for(random_mesh.validate_routes, delay=6, num_sec=30)
```

This code changes to the following:

```python
random_mesh.settle()
```