Implement Node.pid and Node.pgid
The methods return the PID and PGID of a node, respectively. Both
methods can fail, and only work correctly when called on a running node.

Of special note is that `Node.stop` now deletes an entry from the
(thread-unsafe!) module-wide `procs` dict. This makes it harder for
clients to get known-stale PIDs and PGIDs.