Drop unused regex (#22)
As a general rule, unused code should be dropped. If we want this regex
back, it'll be right here in the commit history, and IPv4 regexes are
easy enough to write anyway.

Depending upon the use case, other tools might be more suitable. For
example, if pulling the IP address out of a URL, it would be more robust
to use the `urllib.parse.urlparse` and `ipaddress` modules from the
stdlib.