Rename Node.find_controller to Node.controllers
The method is named in a misleading way. It implies that exactly one
controller node will be returned. But that's not the case, and in
reality, an iterable of zero or more node objects will be returned.

Rename the method to more correctly communicate what it actually does.

Also, make the method return a tuple instead of a list. Immutable data
types are generally safer.