Do nothing when Node.stop() called on dead node
If `Node.stop()` is called on a dead node, it will throw an exception.
Change the method by making it idempotent. Do nothing if the node is
already dead.

Making this change fixes some failed tests where the body of the test
stops a node, and then as a clean-up action, the mesh also tries to stop
the same node.